Remove duplication from autoInstall* profiles
This works on the principle that downstream projects
include the relevant plugin.

The configuration that enables installation is set in
a profile in the parent pom, but only in a pliginManagement
section. Thus the profile does not switch the plugin on/off
but it instead switches the configuration on/off. This way
profiles for different sub-project types (i.e. bundle and
content package) can be run on the full build, as e.g. the
contentpackage plugin is not enabled in a bundle sub-project.