Update AEM Forms solutions with new names. (#824)
Going forward, AEM Forms will have two different solutions on
AEM as a Cloud Service. These are named:
 - formsenrollment
 - formscommunications

As per the existing contract with Cloud Manager during VSTS repository
generation, the updated solution names will be passed instead of the
earler used `includeForms` in the archetype command.

This requires us to update all usage of the `includeForms` property within
the archetype with the updated solution names (`includeFormsenrollment`
and `includeFormscommunications`).

For backward compatibility reasons the older `includeForms` check is
left intact in the archetype template but has been removed from the
README.md.